+++ sickrage/providers/torrent/tokyotoshokan.py
@@ -47,7 +47,7 @@ class TokyoToshokanProvider(TorrentProvider):
     def search(self, search_strings, age=0, ep_obj=None):
         results = []
 
-        if not self.show or not self.show.is_anime:
+        if not ep_obj.show or not ep_obj.show.is_anime:
             return results
 
         for mode in search_strings: